+++ crates/miden-standards/src/account/auth/multisig.rs
@@ -25,7 +25,7 @@ use miden_protocol::utils::sync::LazyLock;
 
 use crate::account::account_component_code;
 
-account_component_code!(MULTISIG_CODE, "auth/multisig.masl");
+account_component_code!(MULTISIG_CODE, "miden-standards-auth-multisig.masp");
 
 // CONSTANTS
 // ================================================================================================