+++ crates/miden-testing/src/kernel_tests/tx/test_asset.rs
@@ -42,7 +42,7 @@ async fn test_create_fungible_asset_succeeds() -> anyhow::Result<()> {
 
     let code = format!(
         "
-        use $kernel::prologue
+        use miden::tx_kernel_core::prologue
         use miden::protocol::faucet
 
         begin
@@ -81,7 +81,7 @@ async fn test_create_non_fungible_asset_succeeds() -> anyhow::Result<()> {
 
     let code = format!(
         "
-        use $kernel::prologue
+        use miden::tx_kernel_core::prologue
         use miden::protocol::faucet
 
         begin
@@ -144,7 +144,7 @@ async fn test_validate_non_fungible_asset(
 ) -> anyhow::Result<()> {
     let code = format!(
         "
-        use $kernel::non_fungible_asset
+        use miden::tx_kernel_core::non_fungible_asset
 
         begin
             # a random asset value
@@ -222,7 +222,7 @@ async fn test_validate_fungible_asset(
 ) -> anyhow::Result<()> {
     let code = format!(
         "
-        use $kernel::fungible_asset
+        use miden::tx_kernel_core::fungible_asset
 
         begin
             push.{ASSET_VALUE}
@@ -276,7 +276,7 @@ async fn test_validate_asset_metadata(
 ) -> anyhow::Result<()> {
     let code = format!(
         "
-        use $kernel::asset
+        use miden::tx_kernel_core::asset
 
         begin
             push.{asset_metadata}
@@ -310,7 +310,7 @@ async fn test_key_to_asset_metadata(
 
     let code = format!(
         "
-        use $kernel::asset
+        use miden::tx_kernel_core::asset
 
         begin
             push.{ASSET_KEY}