+++ crates/miden-testing/src/kernel_tests/tx/test_tx.rs
@@ -154,7 +154,7 @@ async fn test_block_procedures() -> anyhow::Result<()> {
 
     let code = "
         use miden::protocol::tx
-        use $kernel::prologue
+        use miden::tx_kernel_core::prologue
 
         begin
             exec.prologue::prepare_transaction
@@ -778,7 +778,7 @@ async fn test_get_script_root_with_script() -> anyhow::Result<()> {
     let code = format!(
         r#"
         use miden::protocol::tx
-        use $kernel::prologue
+        use miden::tx_kernel_core::prologue
 
         begin
             exec.prologue::prepare_transaction
@@ -806,7 +806,7 @@ async fn test_get_script_root_with_script() -> anyhow::Result<()> {
 async fn test_get_script_root_without_script() -> anyhow::Result<()> {
     let code = r#"
         use miden::protocol::tx
-        use $kernel::prologue
+        use miden::tx_kernel_core::prologue
 
         begin
             exec.prologue::prepare_transaction